Introduced a new caching method to async and hybrid modes
Introduced three caches to async and hybrid modes for IP, token and
cookie separately. Previously the combination of all three of these
properties were cached and if any other combinations of one or two
properties will get slipped in the async mode. This issue was fixed by
having separate caches for each property.

Signed-off-by: Lakshitha Gunasekara <lakshitha.gunasekara@gmail.com>